Desenvolvimento dos metodos de Aluno Desenvolvimento dos metodos de Prova Configuracao cors e injecoes de depedencias Alteracao entidades do banco de dados e novo dump

+++ Escola/Escola.Dominio/Entidades/Aluno.cs
@@ -6,7 +6,7 @@
 
         public string Nome { get; set; }
 
-        public float Media { get; set; }
+        public double Media { get; set; }
 
         public bool Ativo { get; set; }
     }
